requirements.txt
openai==0.28.1
tiktoken
transformers
torch
pyttsx3
moviepy
PIL
bing_image_downloader
edge_tts
vtt_to_srt3
